Add starting hand to play page
